projects/adrv9001: Add TX Muting Capability
Adds the ability to mute the DAC outputs via the FFH TDD block.
This allows extending the "effective" transition time/shortening
the dwell time when running FFH for configurations where the device and system
clock parameters cant satisfy the timespan needed for the hop window.

The FFH TDD block channel outputs are mapped as follows:
0,1 - Hop signal pulse 1 and 2
2,3 - Tx1 Enable pulse 1 and 2
4,5 - Tx2 Enable pulse 1 and 2
6,7 - Tx1 Mute pulse 1 and 2
8,9 - Tx2 Mute pulse 1 and 2

The FFH TDD block source was changed to be the DAC clock, so the linux device
tree needs to be updated accordingly to use the milliseconds driver properies.

Also routed the SPI port out the spare PMOD on the ZCU102 to use for external
clock generators.

Signed-off-by: Brent Kowal <brent.kowal@analog.com>